+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 * 1.0.0-beta (2015-01-12)
 
  * renamed `SchemaException` to `InvalidSchemaException`
+ * changed `JsonValidator::validate()` to return the discovered errors instead
+   of throwing an exception
 
 * 1.0.0-alpha1 (2014-12-03)